Add xzcat to dd script

+++ dd-img.sh
@@ -9,6 +9,7 @@ if [ -z "$img" ]; then
   set +e
   ls *.iso 2> /dev/null
   ls *.img 2> /dev/null
+  ls *.raw.xz 2> /dev/null
   set -e
   echo
   read -p "Image to write: " img
@@ -24,5 +25,9 @@ if [ -z "$dev" ]; then
   echo
 fi
 
-sudo dd if=$img of=$dev bs=4M status=progress
+if [[ $img == *.raw.xz ]]; then
+  xzcat $img | sudo dd of=$dev bs=4M status=progress
+else
+  sudo dd if=$img of=$dev bs=4M status=progress
+fi
